requirements.txt
pytest == 8.4.1
requests == 2.32.4
allure-python-commons == 2.15.0
selenium == 4.35.0
webdriver-manager ==4.0.2